+++ qrand/helpers/__init__.py
@@ -1,7 +1,7 @@
 ##    _____  _____
 ##   |  __ \|  __ \    AUTHOR: Pedro Rivero
 ##   | |__) | |__) |   ---------------------------------
-##   |  ___/|  _  /    DATE: May 18, 2021
+##   |  ___/|  _  /    DATE: May 26, 2021
 ##   | |    | | \ \    ---------------------------------
 ##   |_|    |_|  \_\   https://github.com/pedrorrivero
 ##
@@ -21,11 +21,21 @@
 ## limitations under the License.
 
 from .bounded_factorization import compute_bounded_factorization
-from .is_bitstring import is_bitstring
-from .reverse_endian import reverse_endian
+from .numeral_encodings import (
+    ALPHABETS,
+    isbitstring,
+    isnumeral,
+    numeral_decode,
+    numeral_encode,
+    reverse_endian,
+)
 
 __all__ = [
     "compute_bounded_factorization",
-    "is_bitstring",
+    "ALPHABETS",
+    "isbitstring",
+    "isnumeral",
+    "numeral_decode",
+    "numeral_encode",
     "reverse_endian",
 ]