+++ qrand/helpers/bounded_factorization.py
@@ -22,6 +22,8 @@
 
 from typing import Tuple
 
+from ..errors import validate_natural_number
+
 
 ###############################################################################
 ## COMPUTE BOUNDED FACTORIZATION
@@ -29,7 +31,7 @@ from typing import Tuple
 def compute_bounded_factorization(
     n: int, bound_A: int, bound_B: int
 ) -> Tuple[int, int]:
-    _validate_arguments(n, bound_A, bound_B)
+    _validate_args(n, bound_A, bound_B)
     if bound_A * bound_B < n:
         return bound_A, bound_B
     swapped: bool = bound_A > bound_B
@@ -49,8 +51,7 @@ def compute_bounded_factorization(
     return (final_b, final_a) if swapped else (final_a, final_b)
 
 
-def _validate_arguments(n: int, bound_A: int, bound_B: int) -> None:
-    if not (type(n) is int and type(bound_A) is int and type(bound_B) is int):
-        raise TypeError("Input arguments must be natural numbers.")
-    if not (n > 0 and bound_A > 0 and bound_B > 0):
-        raise ValueError("Input arguments must be natural numbers.")
+def _validate_args(n: int, bound_A: int, bound_B: int) -> None:
+    validate_natural_number(n, zero=False)
+    validate_natural_number(bound_A, zero=False)
+    validate_natural_number(bound_B, zero=False)