+++ qrand/protocols/hadamard.py
@@ -1,7 +1,7 @@
 ##    _____  _____
 ##   |  __ \|  __ \    AUTHOR: Pedro Rivero
 ##   | |__) | |__) |   ---------------------------------
-##   |  ___/|  _  /    DATE: May 20, 2021
+##   |  ___/|  _  /    DATE: May 24, 2021
 ##   | |    | | \ \    ---------------------------------
 ##   |_|    |_|  \_\   https://github.com/pedrorrivero
 ##
@@ -46,9 +46,7 @@ class HadamardProtocol(BareQuantumProtocol):
     @max_bits.setter
     def max_bits(self, max_bits: Optional[int]) -> None:
         max_bits = (
-            max_bits
-            if max_bits and type(max_bits) is int and max_bits > 0
-            else None
+            max_bits if isinstance(max_bits, int) and max_bits > 0 else None
         )
         self._max_bits: Optional[int] = max_bits
 
@@ -60,8 +58,8 @@ class HadamardProtocol(BareQuantumProtocol):
         job: QuantumJob = factory.create_job(
             circuit, backend, num_measurements
         )
-        output: List[str] = job.execute()
-        return self._parse_output(output)
+        measurements: List[str] = job.execute()
+        return self._parse_measurements(measurements)
 
     def verify(self) -> Literal[False]:
         return False
@@ -74,10 +72,10 @@ class HadamardProtocol(BareQuantumProtocol):
             circuit.measure(q)
 
     @staticmethod
-    def _parse_output(output: List[str]) -> BasicResult:
+    def _parse_measurements(measurements: List[str]) -> BasicResult:
         bitstring: str = ""
-        for measurement in output:
-            bitstring += measurement
+        for m in measurements:
+            bitstring += m
         return BasicResult(bitstring)
 
     def _partition_job(self, backend: QuantumBackend) -> Tuple[int, int]: