+++ qrand/qrngV1.py
@@ -24,15 +24,14 @@ import math
 from struct import pack, unpack
 from typing import Optional
 
-from .errors import raise_future_warning, validate_type
 from .helpers import ALPHABETS, numeral_encode
 from .quantum_bit_generator import QuantumBitGenerator
 
 
 ###############################################################################
-## QRNG (OBJECT WRAPPER)
+## QRNG (CLASS WRAPPER)
 ###############################################################################
-class Qrng:
+class Qrng(QuantumBitGenerator):
     """
     A quantum random number generator.
 
@@ -42,15 +41,7 @@ class Qrng:
 
     Parameters
     ----------
-    quantum_bit_generator: QuantumBitGenerator
-        A QuantumBitGenerator instance object to handle random number
-        production.
-
-    Attributes
-    ----------
-    quantum_bit_generator: QuantumBitGenerator
-        A QuantumBitGenerator instance object to handle random number
-        production.
+    Same as QuantumBitGenerator.
 
     Methods
     -------
@@ -70,10 +61,10 @@ class Qrng:
         Returns a random complex in rectangular form from a given polar range.
         If no max radius give, 1 is used. If no max angle given, 2pi used.
     get_random_complex_rect(
-        r1: float = -1,
-        r2: float = +1,
-        i1: Optional[float] = None,
-        i2: Optional[float] = None,
+        r_max: float = 1,
+        r_min: float = 0,
+        i_max: Optional[float] = None,
+        i_min: Optional[float] = None,
     ) -> complex:
         Returns a random complex with both real and imaginary parts from the
         given ranges. Default real range [-1,1). If no imaginary range
@@ -81,18 +72,18 @@ class Qrng:
     get_random_decimal(num_bits: Optional[int] = None) -> str:
         Returns a random decimal base encoded numeral string from a `num_bits`
         uniform distribution.
-    get_random_double(min: float = -1, max: float = +1) -> float:
+    get_random_double(max: float = 1, min: float = 0) -> float:
         Returns a random double from a uniform distribution in the range
-        [min,max). Default range [-1,1).
-    get_random_float(min: float = -1, max: float = +1) -> float:
+        [min,max). Default range [0,1).
+    get_random_float(max: float = 1, min: float = 0) -> float:
         Returns a random float from a uniform distribution in the range
-        [min,max). Default range [-1,1).
+        [min,max). Default range [0,1).
     get_random_hex(num_bits: Optional[int] = None) -> str:
         Returns a random hex base encoded numeral string from a `num_bits`
         uniform distribution.
-    get_random_int(min: int = -1, max: int = +1) -> int:
+    get_random_int(max: int = 1, min: int = 0) -> int:
         Returns a random integer between and including [min, max]. Default
-        range [-1,1].
+        range [0,1].
     get_random_int32() -> int:
         Returns a random 32 bit unsigned integer from a uniform distribution.
     get_random_int64() -> int:
@@ -100,6 +91,11 @@ class Qrng:
     get_random_octal(num_bits: Optional[int] = None) -> str:
         Returns a random octal base encoded numeral string from a `num_bits`
         uniform distribution.
+    get_random_string(
+        num_bits: Optional[int] = None,
+        base_alphabet: str = ALPHABETS["DEFAULT"],
+    ) -> str:
+        Returns a random string with `num_bits` of entropy.
     get_random_uint(num_bits: Optional[int] = None) -> int:
         Returns a random unsigned int from a `num_bits` uniform distribution.
 
@@ -116,60 +112,14 @@ class Qrng:
         - Rename internal variables
         - Add static type hints
         - Fix endianness for bit manipulation.
+        - Reverse min, max arguments' order
         - Add additional features
     """
 
-    def __init__(self, quantum_bit_generator: QuantumBitGenerator):
-        raise_future_warning("This version of Qrng", "1.0.0")
-        self.quantum_bit_generator = quantum_bit_generator
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
     ############################### PUBLIC API ###############################
-    @property
-    def quantum_bit_generator(self) -> QuantumBitGenerator:
-        """
-        A QuantumBitGenerator instance object to handle random number
-        production.
-        """
-        raise_future_warning("quantum_bit_generator", "1.0.0")
-        return self._quantum_bit_generator
-
-    @quantum_bit_generator.setter
-    def quantum_bit_generator(
-        self, quantum_bit_generator: QuantumBitGenerator
-    ) -> None:
-        raise_future_warning("quantum_bit_generator", "1.0.0")
-        validate_type(quantum_bit_generator, QuantumBitGenerator)
-        self._quantum_bit_generator: QuantumBitGenerator = (
-            quantum_bit_generator
-        )
-
-    @property
-    def state(self) -> dict:
-        """
-        The state of the Qrng object.
-        """
-        raise_future_warning("state", "1.0.0")
-        return {
-            "quantum_bit_generator": self.quantum_bit_generator.state,
-        }
-
-    def get_bit_string(self, num_bits: Optional[int] = None) -> str:
-        """
-        Returns a random bitstring from a `num_bits` uniform distribution.
-
-        Parameters
-        ----------
-        num_bits: int, default: BITS (i.e. 32 or 64)
-            Number of bits to retrieve.
-
-        Returns
-        -------
-        out: str
-            Random bitstring of length `num_bits`.
-        """
-        raise_future_warning("get_bit_string", "1.0.0", "get_random_bitstring")
-        return self.quantum_bit_generator.random_bitstring(num_bits)
-
     def get_random_base32(self, num_bits: Optional[int] = None) -> str:
         """
         Returns a random base32 encoded numeral string from a `num_bits`
@@ -220,7 +170,7 @@ class Qrng:
         out: str
             Random bitstring of length `num_bits`.
         """
-        return self.quantum_bit_generator.random_bitstring(num_bits)
+        return self.random_bitstring(num_bits)
 
     def get_random_bytes(self, num_bytes: Optional[int] = None) -> bytes:
         """
@@ -238,11 +188,11 @@ class Qrng:
         """
         num_bits: int = (
             num_bytes * 8
-            if num_bytes and type(num_bytes) is int and num_bytes > 0
-            else self.quantum_bit_generator.BITS
+            if isinstance(num_bytes, int) and num_bytes > 0
+            else self.BITS
         )
         num_bytes = num_bits // 8
-        uint: int = self.quantum_bit_generator.random_uint(num_bits)
+        uint: int = self.random_uint(num_bits)
         return uint.to_bytes(num_bytes, "big")
 
     def get_random_complex_polar(
@@ -254,9 +204,9 @@ class Qrng:
 
         Parameters
         ----------
-        r: float, default 1
+        r: float, default: 1
             Real lower bound for the random number.
-        theta: float, default 2pi
+        theta: float, default: 2pi
             Real strict upper bound for the random number.
 
         Returns
@@ -264,43 +214,43 @@ class Qrng:
         out: complex
             Random complex in the range [0,r) * exp{ j[0,theta) }.
         """
-        r0: float = r * math.sqrt(self.get_random_double(0, 1))
-        theta0: float = self.get_random_double(0, theta)
+        r0: float = r * math.sqrt(self.random_double(1, 0))
+        theta0: float = self.random_double(theta, 0)
         return r0 * math.cos(theta0) + r0 * math.sin(theta0) * 1j
 
     def get_random_complex_rect(
         self,
-        r1: float = -1,
-        r2: float = +1,
-        i1: Optional[float] = None,
-        i2: Optional[float] = None,
+        r_max: float = 1,
+        r_min: float = 0,
+        i_max: Optional[float] = None,
+        i_min: Optional[float] = None,
     ) -> complex:
         """
         Returns a random complex with both real and imaginary parts from the
-        given ranges. Default real range [-1,1). If no imaginary range
+        given ranges. Default real range [0,1). If no imaginary range
         specified, real range used.
 
         Parameters
         ----------
-        r1: float, default -1
+        r_max: float, default: 1
             Real lower bound for the random number.
-        r2: float, default +1
+        r_min: float, default: 0
             Real strict upper bound for the random number.
-        i1: float, default None
+        i_max: float, default: None
             Imaginary lower bound for the random number.
-        i2: float, default None
+        i_min: float, default: None
             Imaginary strict upper bound for the random number.
 
         Returns
         -------
         out: complex
-            Random complex in the range [r1,r2) + j[i1,i2).
+            Random complex in the range [r_max,r_min) + j[i_max,i_min).
         """
-        re: float = self.get_random_double(r1, r2)
-        if i1 is None or i2 is None:
-            im: float = self.get_random_double(r1, r2)
+        re: float = self.random_double(r_max, r_min)
+        if i_max is None or i_min is None:
+            im: float = self.random_double(r_max, r_min)
         else:
-            im = self.get_random_double(i1, i2)
+            im = self.random_double(i_max, i_min)
         return re + im * 1j
 
     def get_random_decimal(self, num_bits: Optional[int] = None) -> str:
@@ -318,42 +268,39 @@ class Qrng:
         out: str
             Random decimal base encoded numeral string.
         """
-        uint: int = self.quantum_bit_generator.random_uint(num_bits)
+        uint: int = self.random_uint(num_bits)
         return f"{uint:d}"
 
-    def get_random_double(self, min: float = -1, max: float = +1) -> float:
+    def get_random_double(self, max: float = 1, min: float = 0) -> float:
         """
         Returns a random double from a uniform distribution in the range
-        [min,max). Default range [-1,1).
+        [min,max). Default range [0,1).
 
         Parameters
         ----------
-        min: float, default -1
-            Lower bound for the random number.
-        max: float, default +1
+        max: float, default: 1
             Strict upper bound for the random number.
+        min: float, default: 0
+            Lower bound for the random number.
 
         Returns
         -------
         out: float
-            Random float in the range [min,max).
+            Random double in the range [min,max).
         """
-        min, max = float(min), float(max)
-        delta: float = max - min
-        shifted: float = self.quantum_bit_generator.random_double(delta)
-        return shifted + min
+        return self.random_double(max, min)
 
-    def get_random_float(self, min: float = -1, max: float = +1) -> float:
+    def get_random_float(self, max: float = 1, min: float = 0) -> float:
         """
         Returns a random float from a uniform distribution in the range
-        [min,max). Default range [-1,1).
+        [min,max). Default range [0,1).
 
         Parameters
         ----------
-        min: float, default -1
-            Lower bound for the random number.
-        max: float, default +1
+        max: float, default: 1
             Strict upper bound for the random number.
+        min: float, default: 0
+            Lower bound for the random number.
 
         Returns
         -------
@@ -373,9 +320,7 @@ class Qrng:
             point_format&oldid=1024960263 (accessed May 25, 2021).
         """
         min, max = float(min), float(max)
-        bits_as_uint: int = (
-            0x3F800000 | self.quantum_bit_generator.random_uint(32 - 9)
-        )
+        bits_as_uint: int = 0x3F800000 | self.random_uint(32 - 9)
         to_bytes: bytes = pack(">I", bits_as_uint)
         standard_value: float = unpack(">f", to_bytes)[0] - 1.0
         return (max - min) * standard_value + min
@@ -395,20 +340,20 @@ class Qrng:
         out: str
             Random hex base encoded numeral string.
         """
-        uint: int = self.quantum_bit_generator.random_uint(num_bits)
+        uint: int = self.random_uint(num_bits)
         return f"{uint:X}"
 
-    def get_random_int(self, min: int = -1, max: int = +1) -> int:
+    def get_random_int(self, max: int = 1, min: int = 0) -> int:
         """
         Returns a random integer between and including [min, max]. Default
-        range [-1,1].
+        range [0,1].
 
         Parameters
         ----------
-        min: int, default -1
-            Lower bound for the random int.
-        max: int, default +1
-            Upper bound for the random int.
+        max: int, default: 1
+            Upper bound for the random number.
+        min: int, default: 0
+            Lower bound for the random number.
 
         Returns
         -------
@@ -417,9 +362,9 @@ class Qrng:
         """
         delta: int = max - min
         num_bits: int = math.floor(math.log(delta, 2)) + 1
-        shifted: int = self.quantum_bit_generator.random_uint(num_bits)
+        shifted: int = self.random_uint(num_bits)
         while shifted > delta:
-            shifted = self.quantum_bit_generator.random_uint(num_bits)
+            shifted = self.random_uint(num_bits)
         return shifted + min
 
     def get_random_int32(self) -> int:
@@ -431,7 +376,7 @@ class Qrng:
         out: int
             Random 32 bit unsigned int.
         """
-        return self.quantum_bit_generator.random_uint(32)
+        return self.random_uint(32)
 
     def get_random_int64(self) -> int:
         """
@@ -442,7 +387,7 @@ class Qrng:
         out: int
             Random 64 bit unsigned int.
         """
-        return self.quantum_bit_generator.random_uint(64)
+        return self.random_uint(64)
 
     def get_random_octal(self, num_bits: Optional[int] = None) -> str:
         """
@@ -459,8 +404,31 @@ class Qrng:
         out: str
             Random octal base encoded numeral string.
         """
-        uint: int = self.quantum_bit_generator.random_uint(num_bits)
+        uint: int = self.random_uint(num_bits)
         return f"{uint:o}"
+
+    def get_random_string(
+        self,
+        num_bits: Optional[int] = None,
+        base_alphabet: str = ALPHABETS["DEFAULT"],
+    ) -> str:
+        """
+        Returns a random string with `num_bits` of entropy.
+
+        Parameters
+        ----------
+        num_bits: int, default: BITS (i.e. 32 or 64)
+            Number of bits to retrieve.
+        base_alphabet: str, default: `qrand.helpers.ALPHABETS['DEFAULT']`
+            A string containig the base alphabet to be used for the output.
+
+        Returns
+        -------
+        out: str
+            Random string with `num_bits` of entropy.
+        """
+        uint: int = self.random_uint(num_bits)
+        return numeral_encode(uint, base_alphabet)
 
     def get_random_uint(self, num_bits: Optional[int] = None) -> int:
         """
@@ -476,4 +444,4 @@ class Qrng:
         out: int
             Random unsigned int of size `num_bits`.
         """
-        return self.quantum_bit_generator.random_uint(num_bits)
+        return self.random_uint(num_bits)